+++ firmware/boron-ins-fsm/test/consoleFunctionTests.cpp
@@ -6,6 +6,7 @@
 #include "../src/stateMachine.h"
 
 IMDoorID globalDoorID = {0xAA, 0xAA, 0xAA};
+bool usingExternalBLEAntenna;
 unsigned long state3_stillness_timer;
 bool hasDurationAlertBeenSent;
 bool hasStillnessAlertBeenSent;
@@ -559,6 +560,112 @@ SCENARIO("Reset_Stillness_Timer_For_Alerting_Session", "[reset stillness timer f
 
             THEN("The stillness timer value did not change") {
                 REQUIRE(state3_stillness_timer == initial_state3_stillness_timer_value);
+            }
+        }
+    }
+}
+
+SCENARIO("Set BLE Antenna Type", "[ble antenna type]") {
+    GIVEN("An internal antenna type") {
+        usingExternalBLEAntenna = false;
+
+        WHEN("the function is called with 'e'") {
+            int returnFlag = use_external_ble_antenna("e");
+
+            THEN("antenna type should remain the same") {
+                REQUIRE(usingExternalBLEAntenna == false);
+            }
+
+            THEN("the function should return the stored value") {
+                REQUIRE(returnFlag == 0);
+            }
+        }
+
+        WHEN("the function is called with neither 1, 0 or e") {
+            int returnFlag = use_external_ble_antenna("invalid value");
+
+            THEN("antenna type should remain the same") {
+                REQUIRE(usingExternalBLEAntenna == false);
+            }
+
+            THEN("the function should return -1") {
+                REQUIRE(returnFlag == -1);
+            }
+        }
+
+        WHEN("the function is called with '0'") {
+            int returnFlag = use_external_ble_antenna("0");
+
+            THEN("antenna type should remain the same") {
+                REQUIRE(usingExternalBLEAntenna == false);
+            }
+
+            THEN("the function should return 0") {
+                REQUIRE(returnFlag == 0);
+            }
+        }
+
+        WHEN("the functions is called with '1'") {
+            int returnFlag = use_external_ble_antenna("1");
+
+            THEN("antenna type should change to external") {
+                REQUIRE(usingExternalBLEAntenna == true);
+            }
+
+            THEN("the function should return 1") {
+                REQUIRE(returnFlag == 1);
+            }
+        }
+    }
+
+    GIVEN("An external antenna type") {
+        usingExternalBLEAntenna = true;
+
+        WHEN("the function is called with 'e'") {
+            int returnFlag = use_external_ble_antenna("e");
+
+            THEN("antenna type should remain the same") {
+                REQUIRE(usingExternalBLEAntenna == true);
+            }
+
+            THEN("the function should return the stored value") {
+                REQUIRE(returnFlag == 1);
+            }
+        }
+
+        WHEN("the function is called with neither 1, 0 or e") {
+            int returnFlag = use_external_ble_antenna("invalid value");
+
+            THEN("antenna type should remain the same") {
+                REQUIRE(usingExternalBLEAntenna == true);
+            }
+
+            THEN("the function should return -1") {
+                REQUIRE(returnFlag == -1);
+            }
+        }
+
+        WHEN("the function is called with '0'") {
+            int returnFlag = use_external_ble_antenna("0");
+
+            THEN("antenna type should change to internal") {
+                REQUIRE(usingExternalBLEAntenna == false);
+            }
+
+            THEN("the function should return 0") {
+                REQUIRE(returnFlag == 0);
+            }
+        }
+
+        WHEN("the function is called with '1'") {
+            int returnFlag = use_external_ble_antenna("1");
+
+            THEN("antenna type should remain the same") {
+                REQUIRE(usingExternalBLEAntenna == true);
+            }
+
+            THEN("the function should return 1") {
+                REQUIRE(returnFlag == 1);
             }
         }
     }